[Options] moved options to own file and implemented OptionsManager - GameplayConfig is now only a legacy wrapper for some OptionsManager methods
