+++ qrogue/util/config/__init__.py
@@ -1,13 +1,14 @@
 from .config import Config
 from .error_config import ErrorConfig
-from .gameplay_config import CheatConfig, GameplayConfig, Options
+from .gameplay_config import CheatConfig, GameplayConfig
 from .gameplay_config import ExpeditionConfig, MapConfig
 from .gameplay_config import GateType, InstructionConfig, QuantumSimulationConfig, PuzzleConfig, ScoreConfig
 from .grammar_config import MapGrammarConfig, PuzzleGrammarConfig, SaveGrammarConfig
+from .options import OptionsType, Options, OptionsManager
 from .path_config import FileTypes, PathConfig
 from .py_cui_config import PyCuiConfig, PyCuiColors
 from .test_config import TestConfig
 from .visual_config import ColorCode, ColorConfig, HudConfig, PopupConfig, UIConfig, split_text
 
-GameplayConfig.init_options()
+OptionsManager.init()
 UIConfig.init()
